BigQueryConfigクラスをAdminer名前空間で分離
- BigQueryConfigクラスをplugins/drivers/bigquery/BigQueryConfig.phpに分離
- Adminer名前空間を追加してbigquery.php内の呼び出しを統一
- 同一名前空間内でのクラス参照を簡潔な記述に統一
- i06.md #3の指示に従いクラス分離を完了
