+++ devtools/issues/i06.md
@@ -25,24 +25,11 @@
 # 2. cdkの移植
 * 現在、本番環境用に以下のcdk stackが構成されています
 	* https://github.com/volareinc/carmo-cdk/blob/main/py-infra/stacks/adminer_gbq.py
-
 * これを、このワークスペースのdevtools/cdk に移植してください。
 
 
 
 # 3. クラス分離
-* bigquery.php に含まれているクラスを1ファイル1クラスに分離してください。
-
-* 分離されたファイルはbigquery/に配置してください
-	bigquery/AdminerBigQueryCSS.php
-	bigquery/AdminerLoginBigQuery.php
-	bigquery/BigQueryCacheManager.php
-	bigquery/BigQueryConfig.php
-	bigquery/BigQueryConnectionPool.php
-	bigquery/BigQueryUtils.php
-	bigquery/Db.php
-	bigquery/Driver.php
-	bigquery/Result.php
-	bigquery/includes.php   ...すべてをincludeするファイル
-
-* bigquery.php はincludes.phpをincudeしてすべての機能を統合してください。
+* bigquery.php に含まれているクラスをファイルに分割します。
+* 1クラスつづ作業します。今回はBigQueryConfigクラスを分離します。
+* 分離されたファイルはplugins/drivers/bigquery/に配置してください
