+++ plugins/drivers/bigquery.php
@@ -353,241 +353,12 @@ class AdminerBigQueryCSS extends Plugin
 }
 
 if (isset($_GET["bigquery"])) {
+require_once __DIR__ . '/bigquery/BigQueryCacheManager.php';
+require_once __DIR__ . '/bigquery/BigQueryConnectionPool.php';
 	define('Adminer\DRIVER', "bigquery");
 
-	class BigQueryConnectionPool
-	{
-
-		private static $pool = array();
-		private static $maxConnections = 3;
-		private static $usageTimestamps = array();
-		private static $creationTimes = array();
-		static function getConnection($key, $config)
-		{
-			if (isset(self::$pool[$key])) {
-				self::$usageTimestamps[$key] = time();
-				$age = time() - self::$creationTimes[$key];
-				return self::$pool[$key];
-			}
-			if (count(self::$pool) >= self::$maxConnections) {
-				self::evictOldestConnection();
-			}
-			$startTime = microtime(true);
-			$clientConfig = array(
-				'projectId' => $config['projectId'],
-				'location' => $config['location']
-			);
-			if (isset($config['credentialsPath'])) {
-				$clientConfig['keyFilePath'] = $config['credentialsPath'];
-			}
-			$client = new BigQueryClient($clientConfig);
-			$creationTime = microtime(true) - $startTime;
-			self::$pool[$key] = $client;
-			self::$usageTimestamps[$key] = time();
-			self::$creationTimes[$key] = time();
-			return $client;
-		}
-		private static function evictOldestConnection()
-		{
-			if (empty(self::$usageTimestamps)) {
-				return;
-			}
-			$oldestKey = array_keys(self::$usageTimestamps, min(self::$usageTimestamps))[0];
-			unset(self::$pool[$oldestKey]);
-			unset(self::$usageTimestamps[$oldestKey]);
-			unset(self::$creationTimes[$oldestKey]);
-		}
-		function clearPool()
-		{
-			$count = count(self::$pool);
-			self::$pool = array();
-			self::$usageTimestamps = array();
-			self::$creationTimes = array();
-		}
-		function getStats()
-		{
-			$stats = array(
-				'pool_size' => count(self::$pool),
-				'max_size' => self::$maxConnections,
-				'connections' => array()
-			);
-			foreach (array_keys(self::$pool) as $key) {
-				$stats['connections'][] = array(
-					'key' => substr($key, 0, 8) . '...',
-					'age' => time() - self::$creationTimes[$key],
-					'last_used' => time() - self::$usageTimestamps[$key]
-				);
-			}
-			return $stats;
-		}
-
-		static function getConnectionFromPool($key)
-		{
-			if (isset(self::$pool[$key])) {
-				self::$usageTimestamps[$key] = time();
-				return self::$pool[$key];
-			}
-			return null;
-		}
-	}
-	class BigQueryConfig
-	{
-
-		public const TYPE_MAPPING = array(
-			'STRING' => array('type' => 'varchar', 'length' => null),
-			'BYTES' => array('type' => 'varchar', 'length' => null),
-			'INT64' => array('type' => 'bigint', 'length' => null),
-			'INTEGER' => array('type' => 'bigint', 'length' => null),
-			'FLOAT64' => array('type' => 'double', 'length' => null),
-			'FLOAT' => array('type' => 'double', 'length' => null),
-			'NUMERIC' => array('type' => 'decimal', 'length' => null),
-			'BIGNUMERIC' => array('type' => 'decimal', 'length' => null),
-			'BOOLEAN' => array('type' => 'tinyint', 'length' => 1),
-			'BOOL' => array('type' => 'tinyint', 'length' => 1),
-			'DATE' => array('type' => 'date', 'length' => null),
-			'TIME' => array('type' => 'time', 'length' => null),
-			'DATETIME' => array('type' => 'datetime', 'length' => null),
-			'TIMESTAMP' => array('type' => 'timestamp', 'length' => null),
-			'GEOGRAPHY' => array('type' => 'geometry', 'length' => null),
-			'JSON' => array('type' => 'json', 'length' => null),
-			'ARRAY' => array('type' => 'text', 'length' => null),
-			'STRUCT' => array('type' => 'text', 'length' => null),
-			'RECORD' => array('type' => 'text', 'length' => null),
-		);
-		public const DANGEROUS_SQL_PATTERNS = array(
-			'ddl_dml' => '/;\\s*(?:DROP|ALTER|CREATE|DELETE|INSERT|UPDATE|TRUNCATE)\\s+/i',
-			'union_injection' => '/UNION\\s+(?:ALL\\s+)?SELECT/i',
-			'block_comments' => '/\/\*.*?\*\//s',
-			'line_comments' => '/--[^\\r\\n]*/i',
-			'execute_commands' => '/\\b(?:EXEC|EXECUTE|SP_)\\b/i',
-		);
-		public const SUPPORTED_FEATURES = array(
-			'database' => true,
-			'table' => true,
-			'columns' => true,
-			'sql' => true,
-			'view' => true,
-			'materializedview' => true,
-		);
-		public const UNSUPPORTED_FEATURES = array(
-			'foreignkeys' => false,
-			'indexes' => false,
-			'processlist' => false,
-			'kill' => false,
-			'transaction' => false,
-			'comment' => false,
-			'drop_col' => false,
-			'dump' => false,
-			'event' => false,
-			'move_col' => false,
-			'privileges' => false,
-			'procedure' => false,
-			'routine' => false,
-			'sequence' => false,
-			'status' => false,
-			'trigger' => false,
-			'type' => false,
-			'variables' => false,
-			'descidx' => false,
-			'check' => false,
-			'schema' => false,
-		);
-		public const CACHE_CONFIG = array(
-			'credentials_ttl' => 86400,
-			'location_ttl' => 86400,
-			'databases_ttl' => 600,
-			'tables_ttl' => 600,
-			'fields_ttl' => 300,
-			'apcu_shm_size' => '64M',
-			'connection_pool_max' => 3,
-		);
-		static function mapType($bigQueryType)
-		{
-			$baseType = strtoupper(preg_replace('/\\(.*\\)/', '', $bigQueryType));
-			return self::TYPE_MAPPING[$baseType] ?? array('type' => 'text', 'length' => null);
-		}
-		static function isDangerousQuery($query)
-		{
-			foreach (self::DANGEROUS_SQL_PATTERNS as $pattern) {
-				if (preg_match($pattern, $query)) {
-					return true;
-				}
-			}
-			return false;
-		}
-		static function isFeatureSupported($feature)
-		{
-			return self::SUPPORTED_FEATURES[$feature] ??
-				(self::UNSUPPORTED_FEATURES[$feature] ?? false);
-		}
-	}
-	class BigQueryCacheManager
-	{
-
-		private static array $staticCache = array();
-		private static array $cacheTimestamps = array();
-		private static ?bool $apcuAvailable = null;
-		private static function isApcuAvailable()
-		{
-			if (self::$apcuAvailable === null) {
-				self::$apcuAvailable = extension_loaded('apcu')
-					&& function_exists('\apcu_exists')
-					&& function_exists('\apcu_fetch')
-					&& function_exists('\apcu_store');
-			}
-			return self::$apcuAvailable;
-		}
-		static function get($key, $ttl = 300)
-		{
-			if (self::isApcuAvailable()) {
-				return \apcu_fetch($key);
-			}
-			if (
-				isset(self::$staticCache[$key]) &&
-				(time() - (self::$cacheTimestamps[$key] ?? 0)) < $ttl
-			) {
-				return self::$staticCache[$key];
-			}
-			return false;
-		}
-		static function set($key, $value, $ttl = 300)
-		{
-			$success = false;
-			if (self::isApcuAvailable()) {
-				$success = \apcu_store($key, $value, $ttl);
-			}
-			self::$staticCache[$key] = $value;
-			self::$cacheTimestamps[$key] = time();
-			return $success;
-		}
-		static function clear($pattern = null)
-		{
-			if ($pattern === null) {
-				if (self::isApcuAvailable()) {
-					\apcu_clear_cache();
-				}
-				self::$staticCache = array();
-				self::$cacheTimestamps = array();
-			} else {
-				foreach (array_keys(self::$staticCache) as $key) {
-					if (strpos($key, $pattern) !== false) {
-						unset(self::$staticCache[$key]);
-						unset(self::$cacheTimestamps[$key]);
-					}
-				}
-			}
-		}
-		static function getStats()
-		{
-			$apcuInfo = self::isApcuAvailable() ? \apcu_cache_info() : array();
-			return array(
-				'static_cache_size' => count(self::$staticCache),
-				'apcu_available' => self::isApcuAvailable(),
-				'apcu_info' => $apcuInfo,
-				'cache_keys' => array_keys(self::$staticCache)
-			);
-		}
-	}
+require_once __DIR__ . '/bigquery/BigQueryConfig.php';
+	
 	class BigQueryUtils
 	{
